% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotation_table}
\alias{annotation_table}
\title{UCSC refseq table}
\format{A data frame with 16929 rows and 6 variables}
\usage{
annotation_table
}
\value{
a dataframe object with gene annotation from the most commonly used
organizations.
}
\description{
This dataset was exported from UCSC Genome Browser available from
\url{http://research.stlouisfed.org/fred2}.
}
\references{
\url{http://www.stat.berkeley.edu/~brill/Papers/jspifinal.pdf}
}
\keyword{datasets}
