% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_beta2M.R
\name{beta2m_values}
\alias{beta2m_values}
\title{Convert beta values to mvalues}
\usage{
beta2m_values(probe_name, env, save_data = FALSE)
}
\arguments{
\item{probe_name}{A character string containing the probe name desired from
the selected \code{'name'} probes.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_data_base_data_type_tumor_data'}{ or}
\item{'tumor_data_base_data_type_both_data'}{ (for tumor and not tumor
data
in separated matrices).}}}

\item{save_data}{Logical value where \code{TRUE} indicates that the
concatenate and filtered matrix should be saved in local storage. The
default is FALSE.}
}
\value{
the beta values converted in mValues stored inside the determined
environment name.
}
\description{
Convert beta values to mvalues
}
\examples{
library(DOAGDC)

download_gdc(
   data_type = "methylation",
   data_base = "legacy", work_dir = "~/Desktop", tumor = "CHOL",
   platform = "Illumina Human Methylation 450"
)

concatenate_methylation(
    name = "HIF3A",
    data_base = "legacy",
    platform = "Illumina Human Methylation 450",
    tumor = "CHOL",
    work_dir = "~/Desktop"
)

beta2m_values(
   probe_name = "cg16672562",
   save_data = TRUE,
   env = CHOL_LEGACY_methylation_tumor_data
)
}
