% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_beta2M.R
\name{beta2mValues}
\alias{beta2mValues}
\title{Convert beta values to mvalues}
\usage{
beta2mValues(probeName, env, saveData = FALSE)
}
\arguments{
\item{probeName}{A character string containing the probe name desired from
the selected \code{'Name'} probes.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data
in separated matrices).}}}

\item{saveData}{Logical value where \code{TRUE} indicates that the
concatenate and filtered matrix should be saved in local storage. The
default is FALSE.}
}
\value{
the beta values converted in mValues stored inside the determined
environment name.
}
\description{
Convert beta values to mvalues
}
\examples{
\dontrun{
beta2mValues(probeName = "cg16672562", saveData = TRUE, env = "env name without quotes")
}
}
