% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_clinical_terms.R
\name{check_clinical_terms}
\alias{check_clinical_terms}
\title{Search for clinical terms for all available tumors}
\usage{
check_clinical_terms(Name, workDir = "~/Desktop", tumor = "all",
  dataType, group.number = 2, term, dataBase = "legacy",
  term_keyword = NULL, tumor_with_term = 1, p.cutoff = 0.05,
  FDR.cutoff = 0.05, confidence.level = 0.95, Width = 7,
  Height = 7, Res = 300, Unit = "in", image.format = "svg", env,
  cexAxixX = 16, cexAxixY = 16)
}
\arguments{
\item{Name}{A character string indicating which row name will be used in
mclust separation, e.g. "HIF3A" in the gene expression concatenated matrix.}

\item{workDir}{A character string specifying the path to work directory.}

\item{tumor}{A character vector indicating which tumors should be used in the
analysis. The default is "all".}

\item{dataType}{A character string indicating which dataType will be used to
group the patients. Only needed when the environment was not created yet
and the user does not have the \code{term} argument.}

\item{group.number}{Numerical value indicating how many groups should be
generated. This argument is advisable if subcategory number is larger than
5. The default is 2.}

\item{term}{A character string containing the clinical term to be used.}

\item{dataBase}{A character string specifying \code{"GDC"} for GDC Data
Portal or \code{"legacy"} for GDC Legacy Archive.}

\item{term_keyword}{A character string containing a possible fragment of the
term. Used when the specific term is unknown.}

\item{tumor_with_term}{A numerical value indicating the minimum number of
tumors containing a specific term with the keyword used. The default is 1.}

\item{p.cutoff}{Numerical value indicating the maximum value for p-values.
The default is \code{0.05}.}

\item{FDR.cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{confidence.level}{A numerical value containing the confidence interval
to be used. The default is 0.95.}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
300 and Unit = "px"}.}

\item{image.format}{A character string indicating which image format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data in
separated matrices).}}}

\item{cexAxixX, cexAxixY}{A numerical value giving the amount by which labels
in X and Y axis respectively should be modified relative to the default
size. The default value is 16 for both arguments.}
}
\description{
Search for clinical terms for all available tumors
}
\examples{
#searching for terms having the keyword inserted in common 'tumor_with_term = 3'
check_clinical_terms(tumor = c("BRCA", "UCS", "OV"), term_keyword = "year", tumor_with_term = 3)

#searching for all terms available with at least three tumors in common 'tumor_with_term = 3'
check_clinical_terms(tumor = c("BRCA", "UCS", "OV"), term_keyword = "", tumor_with_term = 3)

#using the analysis with a specified term
check_clinical_terms("UCS", "isoform", term = "vital_status")
}
