% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_coexpression.R
\name{co_expression}
\alias{co_expression}
\title{Co-Expression Analyses}
\usage{
co_expression(Data = NULL, workDir, tumor, normalization = TRUE,
  tumorData = TRUE, traitData = NULL, networkType = "unsigned",
  minModuleSize = 15, max_softpower = 20, nthreads = 1,
  MEDissThres = 0.25, Width = 2000, Height = 1500, Res = 300,
  Unit = "px", image.format = "png", env, saveCheckpoints = FALSE,
  loadCheckpoint = FALSE)
}
\arguments{
\item{Data}{Used for external non-log expression data. Matrix or data frame.
This \code{Data} must have patients/sample code as \code{colnames} and
genes as \code{rownames}.}

\item{workDir}{A character string specifying the path to work directory.}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast cancer.}

\item{normalization}{Logical value where \code{TRUE} specify the desire to
work with normalized files only. When FALSE, in the second run, do not
forget to set env argument. This argument is only applyable to gene and
isoform expression data from GDC Legacy Archive. The default is
\code{TRUE}.}

\item{tumorData}{Logical value where \code{TRUE} specifies the desire to work
with tumor tissue files only. When set to FALSE, it creates two matrices,
one containing tumor data and other containing data from not-tumor tissue.
The default is \code{TRUE}.}

\item{traitData}{A character string where "default" indicates that the trait
data to be used is provide by the output of \code{check_clinical_terms}
function. Use "custom" for data inserted by the user after the
\code{table2GDCRtools} function and with the table object named as
\code{trait_data}. This object must have patients/sample code as
\code{rownames} and the trait categories as \code{colnames}. By default
trait data is not used.}

\item{networkType}{A character string indicating which network type should be
used. WGCNA allows: "unsigned", "signed", and "signed hybrid". The default
is "unsigned".}

\item{minModuleSize}{Numerical value specifying the minimum cluster size. The
default is 15.}

\item{max_softpower}{Numerical value indicating the maximum soft thresholding
power for which the scale free topology fit indices are to be calculated.
The default is 20.}

\item{nthreads}{Numerical value indicating how many threads to allow. The
number of threads should not be more than the number of actual
processors/cores. The default is 1.}

\item{MEDissThres}{Numerical value specifying the maximum dendrogram cut
height for module merging qualified by dissimilarity (i.e., 1-correlation).
The default is 0.25.}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{image.format}{A character string indicating which image format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data in
separated matrices).}}}

\item{saveCheckpoints}{Logical value where TRUE indicates that an external
representation of analysis' objects will be saved to file in disk. The
default is FALSE.}

\item{loadCheckpoint}{Logical value where TRUE indicates that the saved
checkpoint will be loaded and the analysis is going to continue from that
point. The default is FALSE.}
}
\description{
Powered by WGCNA
}
\examples{
\dontrun{
co_expression(Data, workDir, tumor, env = "env name without quotes")
}
}
