% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_mutation.R
\name{concatenate_mutation}
\alias{concatenate_mutation}
\title{Concatenate GDC files into a single matrix and prepar the data}
\usage{
concatenate_mutation(
  name,
  data_base,
  work_dir,
  tumor,
  workflow_type,
  tumor_data = TRUE,
  platform = "",
  env,
  save_data = FALSE
)
}
\arguments{
\item{name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}

\item{data_base}{A character string specifying \code{"GDC"} for GDC Data
Portal or \code{"legacy"} for GDC Legacy Archive.}

\item{work_dir}{A character string specifying the path to work directory.}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast
cancer.}

\item{workflow_type}{A character string specifying the workflow type for
mutation data in "gdc". Where: \itemize{\item{"varscan" -}{
VarScan2 Variant Aggregation and Masking} \item{"mutect" -}{
MuTect2 Variant Aggregation and Masking}\item{"muse" -}{ MuSE
Variant Aggregation and Masking}\item{"somaticsniper" -}{
SomaticSniper Variant Aggregation and Masking}\item{"all" means to}{
concatenate all workflows into a single matrix.}}}

\item{tumor_data}{Logical value where \code{TRUE} specifies the desire to
work
with tumor tissue files only. When set to FALSE, it creates two matrices,
one containing tumor data and other containing data from not-tumor tissue.
The default is \code{TRUE}.}

\item{platform}{A character string indicating the platform name for
methylation, exon quantificaton, miRNA, and mutation data. \itemize{
\item{For mutation and exon quantificaton data:}{\code{"Illumina GA",
   "Illumina HiSeq" or "all"}.} \item{For methylation data}{\code{"Illumina
   Human Methylation 450", "Illumina Human Methylation 27" or "all"}.}
\item{For miRNA data:}{\code{"Illumina GA", "Illumina HiSeq",
   "H-miRNA_8x15K" (for GBM tumor), "H-miRNA_8x15Kv2" (for OV tumor), or
   "all"}.} }The default for all data_type cited is \code{"all"} (when
downloading data).}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'<tumor>_<data_base>\emph{mutation_tumor_data'}{ or}
\item{'<tumor>}<data_base>_mutation_both_data'}{ (for tumor and not tumor
data
in separated matrices).}}}

\item{save_data}{Logical value where \code{TRUE} indicates that the
concatenate and filtered matrix should be saved in local storage. The
default is FALSE.}
}
\value{
A matrix with data names in row and patients code in column.
}
\description{
\code{concatenate_mutation} is a function designed to concatenate GDC files into
a single matrix, where the columns stand for patients code and rows stand for
data names.
}
\examples{
library(DOAGDC)

# Concatenating gene expression data into a single matrix
# data already downloaded using the 'download_gdc' function
concatenate_mutation(
    name = "HIF3A",
    workflow_type = "varscan",
    platform = "Illumina GA",
    data_base = "legacy",
    tumor = "CHOL",
    work_dir = "~/Desktop"
)
}
