% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_protein.R
\name{concatenate_protein}
\alias{concatenate_protein}
\title{Concatenate GDC files into a single matrix and prepar the data}
\usage{
concatenate_protein(
  name,
  data_base,
  work_dir,
  tumor,
  tumor_data = TRUE,
  only_filter = FALSE,
  tumor_type = 1,
  normal_type = 11,
  env,
  save_data = FALSE
)
}
\arguments{
\item{name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}

\item{data_base}{A character string specifying \code{"GDC"} for GDC Data
Portal or \code{"legacy"} for GDC Legacy Archive.}

\item{work_dir}{A character string specifying the path to work directory.}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast
cancer.}

\item{tumor_data}{Logical value where \code{TRUE} specifies the desire to
work
with tumor tissue files only. When set to FALSE, it creates two matrices,
one containing tumor data and other containing data from not-tumor tissue.
The default is \code{TRUE}.}

\item{only_filter}{Logical value where \code{TRUE} indicates that the matrix
is already concatenate and the function should choose a different
\code{name}, without concatenate all the files again. The default is
FALSE.}

\item{tumor_type}{Numerical value(s) correspondent to barcode data types:
\itemize{Tumor codes: \item{1: Primary Solid Tumor}\item{2: Recurrent
Solid Tumor}\item{3: Primary Blood Derived Cancer - Peripheral
Blood}\item{4: Recurrent Blood Derived Cancer - Bone Marrow}\item{5:
Additional - New Primary}\item{6: Metastatic}\item{7: Additional
Metastatic}\item{8: Human Tumor Original Cells}\item{9: Primary Blood
Derived Cancer - Bone Marrow}} The default is 1.}

\item{normal_type}{Numerical value(s) correspondent to barcode data types:
\itemize{Normal codes: \item{10: Blood Derived Normal}\item{11: Solid
Tissue Normal}\item{12: Buccal Cell Normal}\item{13: EBV Immortalized
Normal}\item{14: Bone Marrow Normal}\item{15: sample type 15}\item{16-19:
sample type 16}}{ or}\itemize{Control codes: \item{use '20:29' without
quotes}} The default is 11.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'<tumor>_<data_base>\emph{protein_tumor_data'}{ or}
\item{'<tumor>}<data_base>_protein_both_data'}{ (for tumor and not tumor
data
in separated matrices).}}}

\item{save_data}{Logical value where \code{TRUE} indicates that the
concatenate and filtered matrix should be saved in local storage. The
default is FALSE.}
}
\value{
A matrix with data names in row and patients code in column.
}
\description{
\code{concatenate_protein} is a function designed to concatenate GDC files
into a single matrix, where the columns stand for patients code and rows
stand for data names.
}
\examples{
library(DOAGDC)

# Concatenating gene expression data into a single matrix
# data already downloaded using the 'download_gdc' function
concatenate_protein(
    name = "Caspase-8-M-E",
    data_base = "legacy",
    tumor = "CHOL",
    work_dir = "~/Desktop"
)
}
