% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_co_expression.R
\name{cross_co_expression}
\alias{cross_co_expression}
\title{Cross co-expression gene list against differential expression genes list}
\usage{
cross_co_expression(
  pair_name = "G2_over_G1",
  final_data,
  width = 2000,
  height = 2000,
  res = 300,
  unit = "px",
  colors = c("green", "blue", "red"),
  venn_diagram_imagetype = "png",
  work_dir,
  fc_cutoff = 2,
  fdr_cutoff = 0.05,
  env
)
}
\arguments{
\item{pair_name}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{final_data}{A character string indicating the name of which diferential
expression package should be used to get the statistical values in the
final list. The default is "EBSeq".}

\item{width}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{height}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{res}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{unit}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{colors}{A character vector indicating the colors to be used in the
venn diagram. The default is c('green', 'blue', "red").}

\item{venn_diagram_imagetype}{A character string indicating the image_format
(e.g. "tiff", "png" or "svg"). The default is "png".}

\item{work_dir}{A character string specifying the path to work directory.}

\item{fc_cutoff}{Numerical value indicating the maximum value for Fold
Change '(FC). The default is \code{2}.}

\item{fdr_cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_data_base_data_type_tumor_data'}{ or}
\item{'tumor_data_base_data_type_both_data'}{ (for tumor and not tumor
data
in separated matrices).}}}
}
\value{
a list of genes co-expressed and differentially expressed genes
inside the determined environment name for further use.
}
\description{
Cross co-expression gene list against differential expression genes list
}
