% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_crosstable.R
\name{CrossThem}
\alias{CrossThem}
\title{Venn diagram of differential expression genes list}
\usage{
CrossThem(FinalData, n.pack = 3, packageNames, pairName = "G2_over_G1",
  Width = 2000, Height = 2000, Res = 300, Unit = "px",
  Colors = c("green", "blue", "red"), VennDiagram.imagetype = "png",
  workDir, env)
}
\arguments{
\item{FinalData}{A character string indicating the name of which diferential
expression package should be used to get the statistical values in the
final list. The default is "EBSeq".}

\item{n.pack}{A numerical value indicating the number of expression analysis
to be used in venn diagram. It is expected the number 2 or 3. The default
is 3.}

\item{packageNames}{A character vector indicating the names of at least two
diferential expression packages used in previous steps: "DESeq2", "edgeR",
"DESeq2, or "All".}

\item{pairName}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
300 and Unit = "px"}.}

\item{Colors}{A character vector indicating the colors to be used in the venn
diagram. The default is c('green', 'blue', "red").}

\item{VennDiagram.imagetype}{A character string indicating the image format
(e.g. "tiff", "png" or "svg"). The default is "png".}

\item{workDir}{A character string specifying the path to work directory.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data in
separated matrices).}}}
}
\description{
Venn diagram of differential expression genes list
}
\examples{
CrossThem("DESeq2", 2, packageNames = c("edgeR", "DESeq2"), env = "env name without quotes")
}
