% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_EBSeq.R
\name{dea_ebseq}
\alias{dea_ebseq}
\title{Run EBSeq gene Differential Expression Analysis (DEA).}
\usage{
dea_ebseq(
  name,
  work_dir,
  env,
  tumor,
  group_gen,
  clinical_pair,
  pair_name = "G2_over_G1",
  rounds = 7,
  norm_type = "QuantileNorm",
  ebtest_qtrm = 0.75,
  ebtest_qtrm_cut = 10,
  p_cutoff = 0.05,
  fdr_cutoff = 0.05,
  fc_cutoff = 2,
  width = 2000,
  height = 1500,
  res = 300,
  unit = "px",
  image_format = "png",
  bullard_quantile = 0.75
)
}
\arguments{
\item{name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}

\item{work_dir}{A character string specifying the path to work directory.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_data_base_data_type_tumor_data'}{ or}
\item{'tumor_data_base_data_type_both_data'}{ (for tumor and not tumor
data
in separated matrices).}}}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast
cancer.}

\item{group_gen}{A character string of the groups generation function:
\itemize{\item{"mclust" - }{\code{groups_identification_mclust()};}
\item{"CoxHR" - }{\code{groups_identification_coxHR()};} \item{"clinical"
\itemize{
\item }{\code{groups_identification_clinical()}.} }
}}

\item{clinical_pair}{A character string containing one of the group pairs
selected after statistical analysis runned in \code{clinical_terms()}
function.}

\item{pair_name}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{rounds}{Numerical value indicating the number of iterations. It is
recommended to check the Alpha and Beta convergence plots in output and
adjust this value until the hyper-parameter estimations converged. The
default is \code{7}.}

\item{norm_type}{A character string indicating which EBSeq normalization
factors type should be used in the analysis "QuantileNorm" or
"MedianNorm". 'The default is \code{"all"}.}

\item{ebtest_qtrm, ebtest_qtrm_cut}{Numerical value. It is removed from the
analysis genes with ebtest_qtrm th quantile < = ebtest_qtrm_cut. More
details in EBSeq \link{EBTest} page. The default is \code{ebtest_qtrm =
   0.75} and \code{ebtest_qtrm_cut = 10}.}

\item{p_cutoff}{Numerical value indicating the maximum value for p-values.
The default is \code{0.05}.}

\item{fdr_cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{fc_cutoff}{Numerical value indicating the maximum value for Fold
Change '(FC). The default is \code{2}.}

\item{width}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{height}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{res}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{unit}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{bullard_quantile}{Numerical value indicating the quantile for the
Bullard's normalization. The default is \code{0.75}.}
}
\value{
A matrix with DE genes in row and statistical values in columns.
}
\description{
Run EBSeq gene Differential Expression Analysis (DEA).
}
\examples{
library(DOAGDC)

# data already downloaded using the 'download_gdc' function
concatenate_expression("gene",
    name = "HIF3A",
    data_base = "legacy",
    tumor = "CHOL",
    work_dir = "~/Desktop"
)

# separating gene HIF3A expression data patients in two groups
groups_identification_mclust("gene", 2,
    name = "HIF3A",
    modelName = "E",
    env = CHOL_LEGACY_gene_tumor_data,
    tumor = "CHOL"
)

# load not normalized data
concatenate_expression("gene",
    normalization = FALSE,
    name = "HIF3A",
    data_base = "legacy",
    tumor = "CHOL",
    env = CHOL_LEGACY_gene_tumor_data,
    work_dir = "~/Desktop"
)

# start DE analysis
# considering concatenate_expression and groups_identification already runned
dea_ebseq(
    pair_name = "G2_over_G1",
    rounds = 7,
    name = "HIF3A",
    env = CHOL_LEGACY_gene_tumor_data
)
}
