% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_reactome.R
\name{DO_React_enrich}
\alias{DO_React_enrich}
\title{DESEASE-ONTOLOGY and REACTOME ENRICHMENT}
\usage{
DO_React_enrich(p_cutoff = 0.05, FDR_cutoff = 0.05, Width = 8,
  Height = 4, Res = 300, Unit = "in", image_format = "png",
  Tool = "edgeR", pairName = "G2_over_G1", env)
}
\arguments{
\item{p_cutoff}{Numerical value indicating the maximum value for p-values.
The default is \code{0.05}.}

\item{FDR_cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{Tool}{A character string indicating which differential expression
analysis tool was last used.}

\item{pairName}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data
in separated matrices).}}}
}
\value{
Enriched terms.
}
\description{
DESEASE-ONTOLOGY and REACTOME ENRICHMENT
}
\examples{
\dontrun{
DO_React_enrich(Tool = "edgeR", env = "env name without quotes")
}
}
