% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_download.gdc.R
\name{download_gdc}
\alias{download_gdc}
\title{Download data from GDC Data Portal and GDC Legacy Archive}
\usage{
download_gdc(
  data_type = "gene",
  tumor,
  data_base = "legacy",
  htseq = "",
  work_dir,
  all_files = FALSE,
  platform = "all"
)
}
\arguments{
\item{data_type}{Type of data. It could be \code{"methylation", "mutation",
   "clinical_supplement", "biospecimen", "gene", or "clinical"(biotab)}.
\itemize{ \item{Only present
in "Legacy" database:}{\code{"protein", "Exon quantification", "miRNA gene
   quantification", "miRNA isoform quantification", "isoform", and "image"}.}
\item{Only present in "GDC" database:}{\code{"miRNA
   Expression Quantification", and "Isoform Expression Quantification"
   (miRNA)}.}}}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast
cancer.}

\item{data_base}{A character string specifying \code{"GDC"} for GDC Data
Portal or \code{"legacy"} for GDC Legacy Archive.}

\item{htseq}{A character string indicating which HTSeq workflow data should
be downloaded: \code{"Counts", "FPKM", or "all"}. The default is
\code{"all"}.}

\item{work_dir}{A character string specifying the path to work directory.}

\item{all_files}{A logical value. Set \code{FALSE} to avoid the download of
not used data to reduce download size, e.g. quantification files. The
default is \code{FALSE}.}

\item{platform}{A character string indicating the platform name for
methylation, exon quantificaton, miRNA, and mutation data. \itemize{
\item{For mutation and exon quantificaton data:}{\code{"Illumina GA",
   "Illumina HiSeq" or "all"}.} \item{For methylation data}{\code{"Illumina
   Human Methylation 450", "Illumina Human Methylation 27" or "all"}.}
\item{For miRNA data:}{\code{"Illumina GA", "Illumina HiSeq",
   "H-miRNA_8x15K" (for GBM tumor), "H-miRNA_8x15Kv2" (for OV tumor), or
   "all"}.} }The default for all data_type cited is \code{"all"} (when
downloading data).}
}
\value{
the files download are stored inside the determined folders in the
user machine.
}
\description{
\code{download_gdc} is a function designed to download methylation, mutation,
clinical data, protein expression, MAGETAB, gene expression, isoform
expression, miRNA expression and clinical images data from GDC Data Portal
and GDC Legacy Archive.
}
\examples{
library(DOAGDC)

# Downloading gene expression data from GDC Legacy Archive
download_gdc("gene", "CHOL", "legacy", work_dir = "~/Desktop")
}
