% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_heatmap_new.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Draw a heatmap}
\usage{
draw_heatmap(
  tool,
  fc_cutoff = 2,
  name,
  method = "euclidean",
  pair_name = "G2_over_G1",
  raw_values = FALSE,
  width = 6,
  height = 6,
  res = 300,
  unit = "in",
  image_format = "svg",
  env,
  scale_method = "row",
  outer_margins = c(0, 0, 0, 0),
  cex_col = 0,
  cex_row = 0,
  degree = 45,
  lab_row = NULL,
  lab_col = NULL
)
}
\arguments{
\item{fc_cutoff}{Numerical value indicating the maximum value for Fold
Change '(FC). The default is \code{2}.}

\item{name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}

\item{method}{The agglomeration method to be used: \code{"euclidean",
   "maximum", "manhattan", "canberra", "binary", "pearson", "abspearson",
   "correlation", "abscorrelation", "spearman" or "kendall"}. The default is
\code{"euclidean"}. More details about \code{method} argument in
\code{amap} package \link{Dist} \code{method} argument.}

\item{pair_name}{A character string indicating the pair name to be used. When
there are only two groups the default is \code{"G2_over_G1"}}

\item{raw_values}{A logical value. If \code{"TRUE"} the expression values are
going to be converted to Z-Score before draw the heat map.}

\item{width}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{height}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{res}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{unit}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_data_base_data_type_tumor_data'}{ or}
\item{'tumor_data_base_data_type_both_data'}{ (for tumor and not tumor
data
in separated matrices).}}}

\item{scale_method}{A character string indicating which method of scale
should
be used: \code{"none"}, \code{"row"}, \code{"column"}. The default is
\code{"row"}. More details about \code{scale_method} argument in
\link{heatmap} \code{scale} argument.}

\item{outer_margins}{A numerical vector of the form c(bottom, left, top,
right) giving the outer margins measured in lines of text. The default is
no outer margins, i.e \code{"c(0,0,0,0)"}. Note: This argument is only
used 'when \code{"lab_row"} and \code{"lab_col"} are set.}

\item{cex_col, cex_row}{A numerical value giving the amount by which
\code{"lab_row"} and \code{"lab_col"} should be modified relative to the
default size.}

\item{degree}{The \code{"lab_col"} rotation in degrees. The default value is
45 degrees.}

\item{lab_row, lab_col}{A logical value. If \code{"TRUE"} it is displayed row
names (\code{"lab_row"}) and col names (\code{"lab_col"}). The default is
\code{"FALSE"} for both, in order to kepp the plot clean.}
}
\value{
a heat map image.
}
\description{
Draw a heatmap
}
\examples{
library(DOAGDC)

# data already downloaded using the 'download_gdc' function
concatenate_expression("gene",
   name = "HIF3A",
   data_base = "legacy",
   tumor = "CHOL",
   work_dir = "~/Desktop"
)

# separating gene HIF3A expression data patients in two groups
groups_identification_mclust("gene", 2,
   name = "HIF3A",
   modelName = "E",
   env = CHOL_LEGACY_gene_tumor_data,
   tumor = "CHOL"
)

# load not normalized data
concatenate_expression("gene",
   normalization = FALSE,
   name = "HIF3A",
   data_base = "legacy",
   tumor = "CHOL",
   env = CHOL_LEGACY_gene_tumor_data,
   work_dir = "~/Desktop"
)

# start DE analysis
# considering concatenate_expression and groups_identification already runned
dea_edger(
   name = "HIF3A",
   group_gen = "mclust",
   env = CHOL_LEGACY_gene_tumor_data
)

draw_heatmap("edgeR", name = "HIF3A", env = CHOL_LEGACY_gene_tumor_data)
}
