% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_annotations}
\alias{GO_annotations}
\title{Gene ontology annotations table}
\format{A list of 3 elements (Biological Process, Cellular Component, and
  Molecular Function)}
\usage{
GO_annotations
}
\description{
This dataset was exported from GO website available from
\url{http://research.stlouisfed.org/fred2}.
}
\references{
\url{http://}
}
\keyword{datasets}
