% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_goest.R
\name{gonto}
\alias{gonto}
\title{Perform Gene-Ontology Pathways Enrichment}
\usage{
gonto(condition, use_genes_without_cat = TRUE, fdr_cutoff = 0.05,
  width = 2000, height = 2000, res = 300, unit = "px",
  image_format = "png", tool, id = "geneID",
  pair_name = "G2_over_G1", env)
}
\arguments{
\item{condition}{A character string containing which condition should be
used: "Upregulated", "Downregulated" or "all".}

\item{use_genes_without_cat}{Logical value where \code{FALSE} indicate that
genes outside the category being tested will be ignored in the calculation
of p-values. The default is "FALSE".}

\item{fdr_cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{width}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{height}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{res}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{unit}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{tool}{A character string indicating which differential expression
analysis tool was last used.}

\item{id}{A character string indicating which id should be used: "hugo",
"gene_symbol", "ensembl" , "refGene" or "geneid". The default is
\code{"geneid"}.}

\item{pair_name}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_data_base_data_type_tumor_data'}{ or}
\item{'tumor_data_base_data_type_both_data'}{ (for tumor and not tumor
data
in separated matrices).}}}
}
\value{
Enriched terms.
}
\description{
Perform Gene-Ontology Pathways Enrichment
}
\examples{
# data already downloaded using the 'download_gdc' function
concatenate_exon("gene",
    name = "HIF3A",
    data_base = "legacy",
    tumor = "CHOL",
    work_dir = "~/Desktop"
)

# separating gene HIF3A expression data patients in two groups
groups_identification_mclust("gene", 2,
    name = "HIF3A",
    modelName = "E",
    env = CHOL_LEGACY_gene_tumor_data,
    tumor = "CHOL"
)

# load not normalized data
concatenate_exon("gene",
    normalization = FALSE,
    name = "HIF3A",
    data_base = "legacy",
    tumor = "CHOL",
    env = CHOL_LEGACY_gene_tumor_data,
    work_dir = "~/Desktop"
)

# start DE analysis
# considering concatenate_exon and groups_identification already runned
dea_edger(
    name = "HIF3A",
    group_gen = "mclust",
    env = CHOL_LEGACY_gene_tumor_data
)

gonto(
    condition = "Upregulated",
    tool = "edgeR", env = CHOL_LEGACY_gene_tumor_data
)
}
