% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_coxph.R
\name{groups_identification_cox_hr}
\alias{groups_identification_cox_hr}
\title{Separate patients in groups}
\usage{
groups_identification_cox_hr(
  name,
  data_type,
  width = 3000,
  height = 2000,
  res = 400,
  unit = "px",
  image_format = "png",
  save_data = TRUE,
  env,
  tumor,
  data_base,
  work_dir
)
}
\arguments{
\item{name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}

\item{data_type}{Type of data. It could be \code{"methylation", "mutation",
   "clinical_supplement", "biospecimen", "gene", or "clinical"(biotab)}.
\itemize{ \item{Only present
in "Legacy" database:}{\code{"protein", "Exon quantification", "miRNA gene
   quantification", "miRNA isoform quantification", "isoform", and "image"}.}
\item{Only present in "GDC" database:}{\code{"miRNA
   Expression Quantification", and "Isoform Expression Quantification"
   (miRNA)}.}}}

\item{width}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{height}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{res}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{unit}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{save_data}{Logical value where \code{TRUE} indicates that the
concatenate and filtered matrix should be saved in local storage. The
default is FALSE.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_data_base_data_type_tumor_data'}{ or}
\item{'tumor_data_base_data_type_both_data'}{ (for tumor and not tumor
data
in separated matrices).}}}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast
cancer.}

\item{data_base}{A character string specifying \code{"GDC"} for GDC Data
Portal or \code{"legacy"} for GDC Legacy Archive.}

\item{work_dir}{A character string specifying the path to work directory.}
}
\value{
the groups generated after using the coxHR analysis.
}
\description{
\code{groups_identification_cox} is a function designed to separate patients
in groups, based in clinical data, and coxHR.
}
\examples{
library(DOAGDC)

concatenate_expression("gene",
    name = "HIF3A",
    data_base = "legacy",
    tumor = "CHOL",
    work_dir = "~/Desktop"
)

groups_identification_cox_hr("HIF3A", "gene",
    tumor = "CHOL",
    work_dir = "~/Desktop",
    data_base = "legacy",
    env = CHOL_LEGACY_gene_tumor_data
)
}
