% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_coxph.R
\name{groups_identification_coxHR}
\alias{groups_identification_coxHR}
\title{Separate patients in groups}
\usage{
groups_identification_coxHR(Name, dataType, Width = 3000,
  Height = 2000, Res = 400, Unit = "px", image.format = "png",
  saveData = TRUE, env, tumor, dataBase, workDir)
}
\arguments{
\item{Name}{A character string indicating which row name will be used in
mclust separation, e.g. "HIF3A" in the gene expression concatenated matrix.}

\item{dataType}{Type of data. It could be \code{"methylation", "mutation",
  "clinical_supplement", "biospecimen", "gene"}. \itemize{ \item{Only present
in "Legacy" database:}{\code{"protein", "Exon quantification", "miRNA gene
  quantification", "miRNA isoform quantification", "isoform", "image",
  "clinical"}.} \item{Only present in "GDC" database:}{\code{"miRNA
  Expression Quantification", "Isoform Expression Quantification"(miRNA)}.} }}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{image.format}{A character string indicating which image format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{saveData}{Logical value where \code{TRUE} indicates that the
concatenate and filtered matrix should be saved in local storage. The
default is FALSE.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data
in separated matrices).}}}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast cancer.}

\item{dataBase}{A character string specifying \code{"GDC"} for GDC Data
Portal or \code{"legacy"} for GDC Legacy Archive.}

\item{workDir}{A character string specifying the path to work directory.}
}
\description{
\code{groups_identification_cox} is a function designed to separate patients in
groups, based in clinical data, and coxHR.
}
\examples{
\dontrun{
groups_identification_coxHR("HIF3A", "gene", tumor = "UCS", dataBase = "legacy")
}
}
