% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg.sets.hs}
\alias{kegg.sets.hs}
\title{Common gene set data collections from \code{gageData} package}
\format{A list of 229 elements (each one correspond to a KEGG Pathway)}
\usage{
kegg.sets.hs
}
\description{
This dataset has 229 KEGG Pathway entries and it is available from
\code{gageData} package.
}
\references{
\code{gageData}
}
\keyword{datasets}
