% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PCA.Analysis.R
\name{PCA_Analysis}
\alias{PCA_Analysis}
\title{Generate plot for Principal Component Analysis}
\usage{
PCA_Analysis(Tool, dataType, Name, workDir, pairName = "G2_over_G1",
  Width = 4, Height = 2, Res = 300, Unit = "in",
  image.format = "png", env)
}
\arguments{
\item{Tool}{A character string indicating which differential expression
analysis tool was last used.}

\item{dataType}{Type of data. It could be \code{"methylation", "mutation",
  "clinical_supplement", "biospecimen", "gene"}. \itemize{ \item{Only present
in "Legacy" database:}{\code{"protein", "Exon quantification", "miRNA gene
  quantification", "miRNA isoform quantification", "isoform", "image",
  "clinical"}.} \item{Only present in "GDC" database:}{\code{"miRNA
  Expression Quantification", "Isoform Expression Quantification"(miRNA)}.} }}

\item{Name}{A character string indicating which row name will be used in
mclust separation, e.g. "HIF3A" in the gene expression concatenated matrix.}

\item{workDir}{A character string specifying the path to work directory.}

\item{pairName}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{image.format}{A character string indicating which image format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data
in separated matrices).}}}
}
\description{
Generate plot for Principal Component Analysis
}
\examples{
\dontrun{
PCA_Analysis("EBSeq", "gene", "HIF3A", pairName = "G2_over_G1", env = "env name without quotes")
}
}
