% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigmet.idx.hs}
\alias{sigmet.idx.hs}
\title{Vector data collections from \code{gageData} package}
\format{A vector with 177 integers.}
\usage{
sigmet.idx.hs
}
\description{
This dataset has 177 values and it is available from
\code{gageData} package.
}
\references{
\code{gageData}
}
\keyword{datasets}
