% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_prepar2my_pack.R
\name{table2GDCRtools}
\alias{table2GDCRtools}
\title{From global to GDCRtools environment}
\usage{
table2GDCRtools(table, dataType, normalization = TRUE, dataBase, tumor,
  tumorData = TRUE, env)
}
\arguments{
\item{table}{A matrix or data frame with data used in this package (i.e Gene,
isoform and protein expression, methylation and mutation data)}

\item{dataType}{Type of data. It could be \code{"methylation", "mutation",
  "clinical_supplement", "biospecimen", "gene"}. \itemize{ \item{Only present
in "Legacy" database:}{\code{"protein", "Exon quantification", "miRNA gene
  quantification", "miRNA isoform quantification", "isoform", "image",
  "clinical"}.} \item{Only present in "GDC" database:}{\code{"miRNA
  Expression Quantification", "Isoform Expression Quantification"(miRNA)}.} }}

\item{normalization}{Logical value where \code{TRUE} specify the desire to
work with normalized files only. When FALSE, in the second run, do not
forget to set env argument. This argument is only applyable to gene and
isoform expression data from GDC Legacy Archive. The default is
\code{TRUE}.}

\item{dataBase}{If the data was downloaded from GDC/Legacy data base, however
not using GDCRtools, please specified which data base. If the data do not
come from GDC/Legacy data base, and it is related with genome version GRCh
37, please insert "legacy" in this argument. Otherwise, if it is related
with genome version GRCh 38, please insert "GDC" in this argument.}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast cancer.}

\item{tumorData}{Logical value where \code{TRUE} specifies the desire to work
with tumor tissue files only. When set to FALSE, it creates two matrices,
one containing tumor data and other containing data from not-tumor tissue.
The default is \code{TRUE}.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data in
separated matrices).}}}
}
\description{
From global to GDCRtools environment
}
\examples{
\dontrun{
patient <- paste0(paste0("patient_", LETTERS[1:4]))
genes <- paste0("gene_", seq(1, 5))
# generate a simulated gene expression matrix
example_gene_table <- matrix(runif(20, 0.0, 90.5), 5, 4, TRUE, list(genes, patient))
# without env created
table2GDCRtools(example_gene_table, dataType = "gene", dataBase = "legacy", tumor = "UCS")
# with env created
table2GDCRtools(example_gene_table, env = "env name without quotes")
}
}
