% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_co_expression.R
\name{cross_co_expression}
\alias{cross_co_expression}
\title{Cross co-expression gene list against differential expression genes list}
\usage{
cross_co_expression(pairName = "G2_over_G1", FinalData, Width = 2000,
  Height = 2000, Res = 300, Unit = "px", Colors = c("green",
  "blue", "red"), VennDiagram_imagetype = "png", workDir,
  FC_cutoff = 2, FDR_cutoff = 0.05, env)
}
\arguments{
\item{pairName}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{FinalData}{A character string indicating the name of which diferential
expression package should be used to get the statistical values in the
final list. The default is "EBSeq".}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{Colors}{A character vector indicating the colors to be used in the
venn diagram. The default is c('green', 'blue', "red").}

\item{VennDiagram_imagetype}{A character string indicating the image_format
(e.g. "tiff", "png" or "svg"). The default is "png".}

\item{workDir}{A character string specifying the path to work directory.}

\item{FC_cutoff}{Numerical value indicating the maximum value for Fold
Change '(FC). The default is \code{2}.}

\item{FDR_cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data
in separated matrices).}}}
}
\value{
a list of genes co-expressed and differentially expressed genes
inside the determined environment name for further use.
}
\description{
Cross co-expression gene list against differential expression genes list
}
\examples{
cross_co_expression(pairName = "G2_over_G1", env = "env name without quotes")

}
