% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_DESeq2.R
\name{dea_deseq2}
\alias{dea_deseq2}
\title{Run DESeq2 gene Differential Expression Analysis (DEA).}
\usage{
dea_deseq2(
  name,
  core_number = 2,
  test = "Default Test",
  group_gen,
  clinical_pair,
  fc_cutoff = 2,
  work_dir,
  tumor,
  fdr_cutoff = 0.05,
  width = 2000,
  height = 1500,
  res = 300,
  unit = "px",
  image_format = "png",
  env,
  cooks_cutoff = FALSE
)
}
\arguments{
\item{name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}

\item{core_number}{A numeric value indicating how many CPU cores should be
used in the analysis. The default value is 2.}

\item{test}{A character string indicating which test should be used:
\code{"LRT", "wald" or "Default Test"}. The default is \code{"Default
   Test"}.}

\item{group_gen}{A character string of the groups generation function:
\itemize{\item{"mclust" - }{\code{groups_identification_mclust()};}
\item{"CoxHR" - }{\code{groups_identification_coxHR()};} \item{"clinical"
\itemize{
\item }{\code{groups_identification_clinical()}.} }
}}

\item{clinical_pair}{A character string containing one of the group pairs
selected after statistical analysis runned in \code{clinical_terms()}
function.}

\item{fc_cutoff}{Numerical value indicating the maximum value for Fold
Change '(FC). The default is \code{2}.}

\item{work_dir}{A character string specifying the path to work directory.}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast
cancer.}

\item{fdr_cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{width}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{height}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{res}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{unit}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_data_base_data_type_tumor_data'}{ or}
\item{'tumor_data_base_data_type_both_data'}{ (for tumor and not tumor
data
in separated matrices).}}}

\item{cooks_cutoff}{Cooks distance remove outliers from the analysis; More
details in DESeq2 \link{results} page. The default is \code{FALSE}.}
}
\value{
A matrix with DE genes in row and statistical values in columns.
}
\description{
Run DESeq2 gene Differential Expression Analysis (DEA).
}
\examples{
library(DOAGDC)

# data already downloaded using the 'download_gdc' function
concatenate_expression("gene",
    name = "HIF3A",
    data_base = "legacy",
    tumor = "CHOL",
    work_dir = "~/Desktop"
)

# separating gene HIF3A expression data patients in two groups
groups_identification_mclust("gene", 2,
    name = "HIF3A",
    modelName = "E",
    env = CHOL_LEGACY_gene_tumor_data,
    tumor = "CHOL"
)

# load not normalized data
concatenate_expression("gene",
    normalization = FALSE,
    name = "HIF3A",
    data_base = "legacy",
    tumor = "CHOL",
    env = CHOL_LEGACY_gene_tumor_data,
    work_dir = "~/Desktop"
)

# start DE analysis
dea_deseq2(
    name = "HIF3A",
    test = "LRT",
    env = CHOL_LEGACY_gene_tumor_data,
    group_gen = "mclust"
)
}
