% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_EBSeq.R
\name{dea_EBSeq}
\alias{dea_EBSeq}
\title{Run EBSeq gene Differential Expression Analysis (DEA).}
\usage{
dea_EBSeq(Name, workDir, env, tumor, groupGen, clinical_pair,
  pairName = "G2_over_G1", rounds = 7, normType = "QuantileNorm",
  EBTest_Qtrm = 0.75, EBTest_QtrmCut = 10, p_cutoff = 0.05,
  FDR_cutoff = 0.05, FC_cutoff = 2, Width = 2000, Height = 1500,
  Res = 300, Unit = "px", image_format = "png",
  Bullard_quantile = 0.75)
}
\arguments{
\item{Name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}

\item{workDir}{A character string specifying the path to work directory.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data
in separated matrices).}}}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast
cancer.}

\item{groupGen}{A character string indicating the groups generation function:
\itemize{\item{"mclust" - }{\code{groups_identification_mclust()};}
\item{"CoxHR" - }{\code{groups_identification_coxHR()};} \item{"clinical"
\itemize{
\item }{\code{groups_identification_clinical()}.} }
}}

\item{clinical_pair}{A character string containing one of the group pairs
selected after statistical analysis runned in \code{clinical_terms()}
function.}

\item{pairName}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{rounds}{Numerical value indicating the number of iterations. It is
recommended to check the Alpha and Beta convergence plots in output and
adjust this value until the hyper-parameter estimations converged. The
default is \code{7}.}

\item{normType}{A character string indicating which EBSeq normalization
factors type should be used in the analysis "QuantileNorm" or
"MedianNorm". 'The default is \code{"all"}.}

\item{EBTest_Qtrm, EBTest_QtrmCut}{Numerical value. It is removed from the
analysis genes with EBTest_Qtrm th quantile < = EBTest_QtrmCut. More
details in EBSeq \link{EBTest} page. The default is \code{EBTest_Qtrm =
   0.75} and \code{EBTest_QtrmCut = 10}.}

\item{p_cutoff}{Numerical value indicating the maximum value for p-values.
The default is \code{0.05}.}

\item{FDR_cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{FC_cutoff}{Numerical value indicating the maximum value for Fold
Change '(FC). The default is \code{2}.}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{Bullard_quantile}{Numerical value indicating the quantile for the
Bullard's normalization. The default is \code{0.75}.}
}
\value{
A matrix with DE genes in row and statistical values in columns.
}
\description{
Run EBSeq gene Differential Expression Analysis (DEA).
}
\examples{
# data already downloaded using the 'download_gdc' function
concatenate_files("gene",
                   Name = "HIF3A",
                   dataBase = "legacy",
                   tumor = "CHOL",
                   workDir = "~/Desktop")

# separating gene HIF3A expression data patients in two groups
groups_identification_mclust("gene", 2,
                               Name = "HIF3A",
                               modelName = "E",
                               env = CHOL_LEGACY_gene_tumor_data,
                               tumor = "CHOL")

# load not normalized data
concatenate_files("gene",
                   normalization = FALSE,
                   Name = "HIF3A",
                   dataBase = "legacy",
                   tumor = "CHOL",
                   env = CHOL_LEGACY_gene_tumor_data,
                   workDir = "~/Desktop")

# start DE analysis
#considering concatenate_files and groups_identification already runned
dea_EBSeq(pairName = "G2_over_G1",
           rounds = 7,
           Name = "HIF3A",
           env = CHOL_LEGACY_gene_tumor_data)
}
