% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_EBSeq.R
\name{dea_EBSeq}
\alias{dea_EBSeq}
\title{Run EBSeq gene Differential Expression Analysis (DEA).}
\usage{
dea_EBSeq(dataType, workDir, Name, env, tumor, groupGen, clinical_pair,
  pairName = "G2_over_G1", rounds = 7, normType = "QuantileNorm",
  EBTest.Qtrm = 0.75, EBTest.QtrmCut = 10, p.cutoff = 0.05,
  FDR.cutoff = 0.05, FC.cutoff = 2, Width = 2000, Height = 1500,
  Res = 300, Unit = "px", image.format = "png",
  Bullard.quantile = 0.75)
}
\arguments{
\item{dataType}{Type of data. It could be \code{"gene"} or \code{"isoform"}.}

\item{workDir}{A character string specifying the path to work directory.}

\item{Name}{A character string indicating which row name will be used in
mclust separation, e.g. "HIF3A" in the gene expression concatenated matrix.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data in
separated matrices).}}}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast cancer.}

\item{groupGen}{A character string indicating the groups generation function:
\itemize{\item{"mclust" - }{\code{groups_identification_mclust()};}
\item{"CoxHR" - }{\code{groups_identification_coxHR()};} \item{"clinical" -
}{\code{groups_identification_clinical()}.} }}

\item{clinical_pair}{A character string containing one of the groups selected
after statistical analysis runned in \code{check_clinical_terms()}
function.}

\item{pairName}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{rounds}{Numerical value indicating the number of iterations. It is
recommended to check the Alpha and Beta convergence plots in output and
adjust this value until the hyper-parameter estimations converged. The
default is \code{7}.}

\item{normType}{A character string indicating which EBSeq normalization
factors type should be used in the analysis "QuantileNorm" or "MedianNorm".
The default is \code{"all"}.}

\item{EBTest.Qtrm, EBTest.QtrmCut}{Numerical value. It is removed from the
analysis genes with EBTest.Qtrm th quantile < = EBTest.QtrmCut. More
details in EBSeq \link{EBTest} page. The default is \code{EBTest.Qtrm =
  0.75} and \code{EBTest.QtrmCut = 10}.}

\item{p.cutoff}{Numerical value indicating the maximum value for p-values.
The default is \code{0.05}.}

\item{FDR.cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{FC.cutoff}{Numerical value indicating the maximum value for Fold Change
(FC). The default is \code{2}.}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{image.format}{A character string indicating which image format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{Bullard.quantile}{Numerical value indicating the quantile for the
Bullard's normalization. The default is \code{0.75}.}
}
\value{
A matrix with DE genes in row and statistical values in columns.
}
\description{
Run EBSeq gene Differential Expression Analysis (DEA).
}
\examples{
\dontrun{
#considering concatenate_files and groups_identification already runned
dea_EBSeq("gene", pairName = "G2_over_G1", rounds = 7, Name = "HIF3A", env = "env_name")
}
}
