% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_edgeR.R
\name{dea_edgeR}
\alias{dea_edgeR}
\title{Run edgeR gene Differential Expression Analysis (DEA).}
\usage{
dea_edgeR(Name, Method = "exacttest", clinical_pair, groupGen,
  FC_cutoff = 2, workDir, env, FDR_cutoff = 0.05, Width = 2000,
  Height = 1500, Res = 300, Unit = "px", image_format = "png")
}
\arguments{
\item{Name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}

\item{Method}{A character string indicating which method should be used:
\code{"exacttest"} or \code{"glmlrt"}. The default is \code{"exacttest"}.}

\item{clinical_pair}{A character string containing one of the group pairs
selected after statistical analysis runned in \code{clinical_terms()}
function.}

\item{groupGen}{A character string indicating the groups generation function:
\itemize{\item{"mclust" - }{\code{groups_identification_mclust()};}
\item{"CoxHR" - }{\code{groups_identification_coxHR()};} \item{"clinical"
\itemize{
\item }{\code{groups_identification_clinical()}.} }
}}

\item{FC_cutoff}{Numerical value indicating the maximum value for Fold
Change '(FC). The default is \code{2}.}

\item{workDir}{A character string specifying the path to work directory.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data
in separated matrices).}}}

\item{FDR_cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
   300 and Unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}
}
\value{
A matrix with DE genes in row and statistical values in columns.
}
\description{
Run edgeR gene Differential Expression Analysis (DEA).
}
\examples{
\dontrun{
#considering concatenate_files and groups_identification already runned
dea_edgeR(pairName = "G2_over_G1", rounds = 7, Name = "HIF3A", env = "env_name")
}
}
