% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_edgeR.R
\name{dea_edger}
\alias{dea_edger}
\title{Run edgeR gene Differential Expression Analysis (DEA).}
\usage{
dea_edger(
  name,
  method = "exacttest",
  clinical_pair,
  group_gen,
  fc_cutoff = 2,
  work_dir,
  env,
  fdr_cutoff = 0.05,
  width = 2000,
  height = 1500,
  res = 300,
  unit = "px",
  image_format = "png"
)
}
\arguments{
\item{name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}

\item{method}{A character string indicating which method should be used:
\code{"exacttest"} or \code{"glmlrt"}. The default is \code{"exacttest"}.}

\item{clinical_pair}{A character string containing one of the group pairs
selected after statistical analysis runned in \code{clinical_terms()}
function.}

\item{group_gen}{A character string of the groups generation function:
\itemize{\item{"mclust" - }{\code{groups_identification_mclust()};}
\item{"CoxHR" - }{\code{groups_identification_coxHR()};} \item{"clinical"
\itemize{
\item }{\code{groups_identification_clinical()}.} }
}}

\item{fc_cutoff}{Numerical value indicating the maximum value for Fold
Change '(FC). The default is \code{2}.}

\item{work_dir}{A character string specifying the path to work directory.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_data_base_data_type_tumor_data'}{ or}
\item{'tumor_data_base_data_type_both_data'}{ (for tumor and not tumor
data
in separated matrices).}}}

\item{fdr_cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{width}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{height}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{res}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{unit}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}
}
\value{
A matrix with DE genes in row and statistical values in columns.
}
\description{
Run edgeR gene Differential Expression Analysis (DEA).
}
\examples{
library(DOAGDC)

# data already downloaded using the 'download_gdc' function
concatenate_expression("gene",
   name = "HIF3A",
   data_base = "legacy",
   tumor = "CHOL",
   work_dir = "~/Desktop"
)

# separating gene HIF3A expression data patients in two groups
groups_identification_mclust("gene", 2,
   name = "HIF3A",
   modelName = "E",
   env = CHOL_LEGACY_gene_tumor_data,
   tumor = "CHOL"
)

# load not normalized data
concatenate_expression("gene",
   normalization = FALSE,
   name = "HIF3A",
   data_base = "legacy",
   tumor = "CHOL",
   env = CHOL_LEGACY_gene_tumor_data,
   work_dir = "~/Desktop"
)

# start DE analysis
dea_edger(
   name = "HIF3A",
   group_gen = "mclust",
   env = CHOL_LEGACY_gene_tumor_data
)
}
