% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_download.gdc.R
\name{download_gdc}
\alias{download_gdc}
\title{Download data from GDC Data Portal and GDC Legacy Archive}
\usage{
download_gdc(dataType = "gene", tumor, dataBase = "legacy",
  HTSeq = "", workDir = "~/Desktop", all.files = FALSE,
  Platform = "all")
}
\arguments{
\item{dataType}{Type of data. It could be \code{"methylation", "mutation",
  "clinical_supplement", "biospecimen", "gene"}. \itemize{ \item{Only present
in "Legacy" database:}{\code{"protein", "Exon quantification", "miRNA gene
  quantification", "miRNA isoform quantification", "isoform", "image",
  "clinical"}.} \item{Only present in "GDC" database:}{\code{"miRNA
  Expression Quantification", "Isoform Expression Quantification"(miRNA)}.} }}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast cancer.}

\item{dataBase}{A character string specifying \code{"GDC"} for GDC Data
Portal or \code{"legacy"} for GDC Legacy Archive.}

\item{HTSeq}{A character string indicating which HTSeq workflow data should
be downloaded: \code{"Counts", "FPKM", "FPKM-UQ" or "all"}. The default is
\code{"all"}.}

\item{workDir}{A character string specifying the path to work directory.}

\item{all.files}{A logical value. Set \code{FALSE} to avoid the download of
not used data to reduce download size, e.g. quantification files. The
default is \code{FALSE}.}

\item{Platform}{A character string indicating the platform name for
methylation, exon quantificaton, miRNA, and mutation data. \itemize{
\item{For mutation and exon quantificaton data:}{\code{"Illumina GA",
  "Illumina HiSeq" or "all"}.} \item{For methylation data}{\code{"Illumina
  Human Methylation 450", "Illumina Human Methylation 27" or "all"}.}
\item{For miRNA data:}{\code{"Illumina GA", "Illumina HiSeq",
  "H-miRNA_8x15K" (for GBM tumor), "H-miRNA_8x15Kv2" (for OV tumor), or
  "all"}.} }The default for all dataType cited is \code{"all"} (when
downloading data).}
}
\description{
\code{download_gdc} is a function designed to download methylation, mutation,
clinical data, protein expression, MAGETAB, gene expression, isoform
expression, miRNA expression and clinical images data from GDC Data Portal
and GDC Legacy Archive.
}
\examples{
\dontrun{
# Downloading gene expression data from GDC Data Portal
download_gdc(dataType = "gene", "BRCA", "GDC", workDir = "~/Desktop", HTSeq = "counts")

# Downloading mutation data from GDC Legacy Archive
download_gdc(dataType = "mutation", "BRCA", "GDC", Platform = "Illumina HiSeq")
}
}
