% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_heatmap_new.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Draw a heatmap}
\usage{
draw_heatmap(Tool, FC.cutoff = 2, dataType, Name, Method = "euclidean",
  pairName = "G2_over_G1", Width = 4, Height = 5, Res = 300,
  Unit = "in", image.format = "svg", env, ScaleMethod = "row",
  outerMargins = c(0, 0, 0, 0), cexCol = 0, cexRow = 0,
  degree = 45, labRow = NULL, labCol = NULL)
}
\arguments{
\item{Tool}{A character string indicating which differential expression
analysis tool was last used.}

\item{FC.cutoff}{Numerical value indicating the maximum value for Fold Change
(FC). The default is \code{2}.}

\item{dataType}{Type of data. It could be \code{"methylation", "mutation",
  "clinical_supplement", "biospecimen", "gene"}. \itemize{ \item{Only present
in "Legacy" database:}{\code{"protein", "Exon quantification", "miRNA gene
  quantification", "miRNA isoform quantification", "isoform", "image",
  "clinical"}.} \item{Only present in "GDC" database:}{\code{"miRNA
  Expression Quantification", "Isoform Expression Quantification"(miRNA)}.} }}

\item{Name}{A character string indicating which row name will be used in
mclust separation, e.g. "HIF3A" in the gene expression concatenated matrix.}

\item{Method}{The agglomeration method to be used: \code{"euclidean",
  "maximum", "manhattan", "canberra", "binary", "pearson", "abspearson",
  "correlation", "abscorrelation", "spearman" or "kendall"}. The default is
\code{"euclidean"}. More details about \code{Method} argument in
\code{amap} package \link{Dist} \code{method} argument.}

\item{pairName}{A character string indicating the pair name to be used. When
there are only two groups the default is \code{"G2_over_G1"}}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{image.format}{A character string indicating which image format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data in
separated matrices).}}}

\item{ScaleMethod}{A character string indicating which method of scale should
be used: \code{"none"}, \code{"row"}, \code{"column"}. The default is
\code{"row"}. More details about \code{ScaleMethod} argument in
\link{heatmap} \code{scale} argument.}

\item{outerMargins}{A numerical vector of the form c(bottom, left, top,
right) giving the outer margins measured in lines of text. The default is
no outer margins, i.e \code{"c(0,0,0,0)"}. Note: This argument is only used
when \code{"labRow"} and \code{"labCol"} are set.}

\item{cexCol, cexRow}{A numerical value giving the amount by which
\code{"labRow"} and \code{"labCol"} should be modified relative to the
default size.}

\item{degree}{The \code{"labCol"} rotation in degrees. The default value is
45 degrees.}

\item{labRow, labCol}{A character vectors with row and column labels desired,
respectively. Another possibility is to assign "rownames" as \code{"labRow"}
and "colnames" \code{"labCol"} as to use the default values from expression
table. The default is none for both, in order to kepp the plot clean.}
}
\description{
Draw a heatmap
}
\examples{
\dontrun{
draw_heatmap("EBSeq", dataType = "gene", Name = "HIF3A", env = "env name without quotes")
}
}
