% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_goest.R
\name{GOnto}
\alias{GOnto}
\title{Perform Gene-Ontology Pathways Enrichment}
\usage{
GOnto(condition, use_genes_without_cat = TRUE, FDR.cutoff = 0.05,
  Width = 2000, Height = 2000, Res = 300, Unit = "px",
  image.format = "png", Tool, ID = "GeneID", pairName = "G2_over_G1",
  env)
}
\arguments{
\item{condition}{A character string containing which condition should be
used: "Upregulated", "Downregulated" or "all".}

\item{use_genes_without_cat}{Logical value where \code{FALSE} indicate that
genes outside the category being tested will be ignored in the calculation
of p-values. The default is "FALSE".}

\item{FDR.cutoff}{Numerical value indicating the maximum value for FDR
values. The default is \code{0.05}.}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{nBreaks = 55, Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{image.format}{A character string indicating which image format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{Tool}{A character string indicating which differential expression
analysis tool was last used.}

\item{ID}{A character string indicating which ID should be used: "HUGO",
"GeneSymbol", "ensembl" , "refGene" or "GeneID". The default is \code{"GeneID"}.}

\item{pairName}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data in
separated matrices).}}}
}
\value{
Enriched terms.
}
\description{
Perform Gene-Ontology Pathways Enrichment
}
\examples{
\dontrun{
GOnto(condition = "Upregulated", Tool = "edgeR", env = "env name without quotes")
}
}
