% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_noMclust.R
\name{groups_identification_mclust}
\alias{groups_identification_mclust}
\title{Separate patients in groups}
\usage{
groups_identification_mclust(dataType, group.number = "",
  modelName = NULL, uncertaintyCutoff = 0.05, reRunPlots = FALSE,
  nBreaks = 55, Width = 2000, Height = 1500, Res = 300,
  Unit = "px", image_format = "png", saveData = TRUE, env, tumor,
  dataBase, workDir = "~/Desktop", Name)
}
\arguments{
\item{dataType}{Type of data. It could be \code{"methylation", "mutation",
  "clinical_supplement", "biospecimen", "gene"}. \itemize{ \item{Only present
in "Legacy" database:}{\code{"protein", "Exon quantification", "miRNA gene
  quantification", "miRNA isoform quantification", "isoform", "image",
  "clinical"}.} \item{Only present in "GDC" database:}{\code{"miRNA
  Expression Quantification", "Isoform Expression Quantification"(miRNA)}.} }}

\item{group.number}{Numerical value indicating how many groups should be
generated.}

\item{modelName}{A character string indicating which mclust modelName will be
used. For more details please check \link{mclustModelNames} help file.}

\item{uncertaintyCutoff}{Numerical value indicating which uncertainty value
for the separation should be tolerated. Patients over this threshold will
be removed from the analysis.}

\item{reRunPlots}{Logical value where TRUE indicate that the function should
run the step of group generation using the \code{uncertaintyCutoff}
parameter for filtering the data. The default is \code{FALSE}.}

\item{nBreaks}{Numerical value giving the number of cells for the \code{hist}
bars. As default \code{nBreaks = 55}.}

\item{Width, Height, Res, Unit}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{saveData}{Logical value where \code{TRUE} indicates that the
concatenate and filtered matrix should be saved in local storage. The
default is FALSE.}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data
in separated matrices).}}}

\item{tumor}{A character string contaning one of the 33 tumors available in
the TCGA project. For instance, the \code{"BRCA"} stands for breast cancer.}

\item{dataBase}{A character string specifying \code{"GDC"} for GDC Data
Portal or \code{"legacy"} for GDC Legacy Archive.}

\item{workDir}{A character string specifying the path to work directory.}

\item{Name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}
}
\description{
\code{groups_identification_mclust} is a function designed to separate patients in
groups, powered by mclust.
}
\examples{
\dontrun{
#separating isoform uc002peh.2 expression data patients in two groups
groups_identification_mclust("isoform", 2, Name = "uc002peh.2", modelName = "E")
}
}
