% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PCA.Analysis.R
\name{PCA_Analysis}
\alias{PCA_Analysis}
\title{Generate plot for Principal Component Analysis}
\usage{
PCA_Analysis(Tool, Name, workDir, pairName = "G2_over_G1", Width = 4,
  Height = 2, Res = 300, Unit = "in", image_format = "png", env)
}
\arguments{
\item{Tool}{A character string indicating which differential expression
analysis tool was last used.}

\item{Name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}

\item{workDir}{A character string specifying the path to work directory.}

\item{pairName}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{Width}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Height}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Res}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{Unit}{Graphical parameters. See \link{par} for more
details. As default \code{Width = 2000, Height = 1500, Res =
  300 and Unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_dataBase_dataType_tumor_data'}{ or}
\item{'tumor_dataBase_dataType_both_data'}{ (for tumor and not tumor data
in separated matrices).}}}
}
\value{
the PCAs plots.
}
\description{
Generate plot for Principal Component Analysis
}
\examples{
\dontrun{
PCA_Analysis("EBSeq", "gene", "HIF3A", pairName = "G2_over_G1", env = "env name without quotes")
}
}
