% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PCA.Analysis.R
\name{pca_analysis}
\alias{pca_analysis}
\title{Generate plot for Principal Component Analysis}
\usage{
pca_analysis(tool, name, work_dir, pair_name = "G2_over_G1", width = 4,
  height = 2, res = 300, unit = "in", image_format = "png", env)
}
\arguments{
\item{tool}{A character string indicating which differential expression
analysis tool was last used.}

\item{name}{A character string indicating the desired values to be used in
next analysis. For instance, "HIF3A" in the legacy gene expression matrix,
"mir-1307" in the miRNA quantification matrix, or "HER2" in the protein
quantification matrix.}

\item{work_dir}{A character string specifying the path to work directory.}

\item{pair_name}{A character string indicating which condition name should be
used. When there are only two groups the default is \code{"G2_over_G1"}.}

\item{width}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{height}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{res}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{unit}{Graphical parameters. See \link{par} for more
details. As default \code{width = 2000, height = 1500, res =
   300 and unit = "px"}.}

\item{image_format}{A character string indicating which image_format will be
used. It could be "png" or "svg". The only unit available in "svg" is
inches ('in'). The default is "png".}

\item{env}{A character string containing the environment name that should be
used. If none has been set yet, the function will create one in global
environment following the standard criteria:
\itemize{\item{'tumor_data_base_data_type_tumor_data'}{ or}
\item{'tumor_data_base_data_type_both_data'}{ (for tumor and not tumor
data
in separated matrices).}}}
}
\value{
the PCAs plots.
}
\description{
Generate plot for Principal Component Analysis
}
\examples{
library(DOAGDC)

# data already downloaded using the 'download_gdc' function
concatenate_expression("gene",
   name = "HIF3A",
   data_base = "legacy",
   tumor = "CHOL",
   work_dir = "~/Desktop"
)

# separating gene HIF3A expression data patients in two groups
groups_identification_mclust("gene", 2,
   name = "HIF3A",
   modelName = "E",
   env = CHOL_LEGACY_gene_tumor_data,
   tumor = "CHOL"
)

# load not normalized data
concatenate_expression("gene",
   normalization = FALSE,
   name = "HIF3A",
   data_base = "legacy",
   tumor = "CHOL",
   env = CHOL_LEGACY_gene_tumor_data,
   work_dir = "~/Desktop"
)

# start DE analysis
# considering concatenate_expression and groups_identification already runned
dea_edger(
   name = "HIF3A",
   group_gen = "mclust",
   env = CHOL_LEGACY_gene_tumor_data
)

pca_analysis("edgeR", "HIF3A", "~/Desktop",
   pair_name = "G2_over_G1",
   env = CHOL_LEGACY_gene_tumor_data
)
}
